+++ plexserviceinterfaces/src/main/java/com/liskovsoft/plexserviceinterfaces/data/PlexMediaItem.java
@@ -1,0 +1,17 @@
+package com.liskovsoft.plexserviceinterfaces.data;
+
+public interface PlexMediaItem {
+    String getRatingKey();
+
+    String getKey();
+
+    String getTitle();
+
+    String getType();
+
+    long getDurationMs();
+
+    String getThumbUrl();
+
+    int getYear();
+}