+++ plexserviceinterfaces/src/main/java/com/liskovsoft/plexserviceinterfaces/data/PlexServer.java
@@ -1,0 +1,13 @@
+package com.liskovsoft.plexserviceinterfaces.data;
+
+public interface PlexServer {
+    String getClientIdentifier();
+
+    String getName();
+
+    String getBaseUrl();
+
+    boolean isOwned();
+
+    boolean isOnline();
+}